+++ src/tests/background/settings.spec.ts
@@ -60,7 +60,7 @@ describe("background/settings", () => {
     const bookImportSettings = await loadBookImportSettings();
 
     expect(windowSettings).toEqual(defaultWindowSettings());
-    expect(usiEngines.engineList).empty;
+    expect(usiEngines.engineList).empty.toBeTruthy();
     expect(appSettings).toEqual(
       defaultAppSettings({
         returnCode: process.platform === "win32" ? "\r\n" : "\n",
